music player styling + move song upload to modal
