+++ frontend/src/index.css
@@ -6,4 +6,5 @@
 body {
     font-family: "Lucida Grande","Lucida Sans Unicode","Lucida Sans","Garuda","Verdana","Tahoma","sans-serif";
     color: #333;
+    padding-bottom: 12.5%;
 }
